Use bestaudio/worst format for youtube-dl
m4a doesn't exist as a format for all sites.
bestaudio/worst will use the best audio quality,
or the worst video quality if no audio format is available.
This does require using -x to make it m4a in the end,
but that's a no-op if bestaudio is m4a already.